+++ src/lib.rs
@@ -5,7 +5,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! evalexpr = "7"
+//! evalexpr = "6"
 //! ```
 //!
 //! Then you can use `evalexpr` to **evaluate expressions** like this: